requirements.txt
ultralytics
opencv-python
numpy
django
django-debug-toolbar==5.1.0